feat: Implement telemetry decoder for Teensy IMU data
- Added `TelemetryDecoder` class to handle binary telemetry stream from Teensy.
- Introduced CRC validation and frame parsing for IMU batches and heartbeat messages.
- Integrated telemetry data publishing to `MeasurementBus`.
- Updated `TeensyManager` to manage telemetry connections and data flow.
- Created unit tests for `TelemetryDecoder` to ensure correct decoding and error handling.
- Added tests for `MeasurementBus` to verify subscriber behavior and message delivery.
- Implemented `ImuConfigRepository` tests to validate configuration persistence and updates.

+++ firmware/src/trigger_engine.h
@@ -30,7 +30,7 @@ constexpr uint32_t kPulseWidthUs = 100;
 
 struct PulseEvent {
     uint32_t idx;     // group-local rising-edge counter, starts at 1 after arm()
-    uint32_t t_us;    // micros() captured at the rising edge
+    uint64_t t_us;    // wrap-extended micros() captured at the rising edge (time64.h)
 };
 
 class TriggerEngine {
